chore: Add local IP to hosts for mobile testing
As my laptop is running on a DHCP server from my router, I have added a
programmatically generated IP address to the allowed hosts. To allow an
extra device to connect to the `runserver` of django, the server must be
started with an optional IP address:

`./manage.py runserver 0.0.0.0:8000`

I have no port mapping through my router so this does not appear to
expose the server to external IP addresses.